refactor rate into its own .hpp file
